Fix `Unexpected token <` error
By default skipWaiting parameter is set to true, allowing newly
registered service worker to bypass the `waiting` state, so all out of
date cache entries from the previous service worker will be deleted.

Because of async dynamic module loading we don't want that to happen,
keeping a way of loading old chunks.

The downside of removing skipWaiting is that newly deployed
service worker code and caches won't become active until all existing
tabs that have an older version of that service worker are closed, but
that actually sounds we want exactly this behaviour.

Related to
https://github.com/GoogleChromeLabs/sw-precache/issues/180#issuecomment-248386760